+++ src/app/jobList/[boardId]/edit/page.tsx
@@ -1,7 +1,8 @@
+import JobBoardWritePage from "@/components/JobBoard-write";
 import React from "react";
 
-const page = () => {
-  return <div>page</div>;
+const JobListEdit = () => {
+  return <JobBoardWritePage isEdit={true} />;
 };
 
-export default page;
+export default JobListEdit;
